+++ pyppeteer/input.py
@@ -183,13 +183,7 @@ class Keyboard(object):
             Modifier keys DO NOT effect :meth:`sendCharacter`. Holding down
             ``shift`` will not type the text in upper case.
         """
-        await self._client.send('Input.dispatchKeyEvent', {
-            'type': 'char',
-            'modifiers': self._modifiers,
-            'text': char,
-            'key': char,
-            'unmodifiedText': char,
-        })
+        await self._client.send('Input.insertText', {'text': char})
 
     async def type(self, text: str, options: Dict = None, **kwargs: Any
                    ) -> None:
@@ -251,7 +245,11 @@ class Keyboard(object):
 
 
 class Mouse(object):
-    """Mouse class."""
+    """Mouse class.
+
+    The :class:`Mouse` operates in main-frame CSS pixels relative to the
+    top-left corner of the viewport.
+    """
 
     def __init__(self, client: CDPSession, keyboard: Keyboard) -> None:
         self._client = client